throw error chat

+++ src/main/java/leare/apiGateway/controllers/graphql/ChatResolver.java
@@ -25,43 +25,43 @@ public class ChatResolver {
     }
 
     @QueryMapping
-    public Chat[] userChats(@ContextValue("Authorization") String AuthorizationHeader) {
+    public Chat[] userChats(@ContextValue("Authorization") String AuthorizationHeader) throws Exception {
 
-        // if (!authConsumer.CheckRoute("/chat/user/:user_id", "get", AuthorizationHeader)) {
-        //     throw new Exception("Auth Problem");
-        // }
+        if (!authConsumer.CheckRoute("/chat/user/:user_id", "get", AuthorizationHeader)) {
+            throw new Exception("Auth Problem");
+        }
 
         String user_id = authConsumer.DecryptToken(AuthorizationHeader).getUserID();
         return chatConsumer.getUserChats(user_id);
     }
 
     @QueryMapping
-    public Message[] chatMessages(@ContextValue("Authorization") String AuthorizationHeader , @Argument String chat_id) {
+    public Message[] chatMessages(@ContextValue("Authorization") String AuthorizationHeader , @Argument String chat_id) throws Exception {
 
-        // if (!authConsumer.CheckRoute("/chat/:chat_id/messages", "get", AuthorizationHeader)) {
-        //     throw new Exception("Auth Problem");
-        // }
+        if (!authConsumer.CheckRoute("/chat/:chat_id/messages", "get", AuthorizationHeader)) {
+            throw new Exception("Auth Problem");
+        }
 
         String user_id = authConsumer.DecryptToken(AuthorizationHeader).getUserID();
         return chatConsumer.getChatMessages(chat_id, user_id);
     }
 
     @MutationMapping
-    public ChatData createChat(@ContextValue("Authorization") String AuthorizationHeader, @Argument ChatInput chat_input) {
+    public ChatData createChat(@ContextValue("Authorization") String AuthorizationHeader, @Argument ChatInput chat_input) throws Exception {
 
-        // if (!authConsumer.CheckRoute("/chat", "post", AuthorizationHeader)) {
-        //     throw new Exception("Auth Problem");
-        // }
+        if (!authConsumer.CheckRoute("/chat", "post", AuthorizationHeader)) {
+            throw new Exception("Auth Problem");
+        }
 
         return chatConsumer.createChat(chat_input);
     }
 
     @MutationMapping
-    public ChatUser joinChat(@ContextValue("Authorization") String AuthorizationHeader,  @Argument String chat_id) {
+    public ChatUser joinChat(@ContextValue("Authorization") String AuthorizationHeader,  @Argument String chat_id) throws Exception {
 
-        // if (!authConsumer.CheckRoute("/chat/:chat_id/join", "patch", AuthorizationHeader)) {
-        //     throw new Exception("Auth Problem");
-        // }
+        if (!authConsumer.CheckRoute("/chat/:chat_id/join", "patch", AuthorizationHeader)) {
+            throw new Exception("Auth Problem");
+        }
 
         String user_id = authConsumer.DecryptToken(AuthorizationHeader).getUserID();
         String user_nickname = authConsumer.DecryptToken(AuthorizationHeader).getUsername();
@@ -69,11 +69,11 @@ public class ChatResolver {
     }
 
     @MutationMapping
-    public Map<String, String> leaveChat(@ContextValue("Authorization") String AuthorizationHeader, @Argument String chat_id) {
+    public Map<String, String> leaveChat(@ContextValue("Authorization") String AuthorizationHeader, @Argument String chat_id) throws Exception {
 
-        // if (!authConsumer.CheckRoute("/chat/:chat_id/leave", "patch", AuthorizationHeader)) {
-        //     throw new Exception("Auth Problem");
-        // }
+        if (!authConsumer.CheckRoute("/chat/:chat_id/leave", "patch", AuthorizationHeader)) {
+            throw new Exception("Auth Problem");
+        }
 
         String user_id = authConsumer.DecryptToken(AuthorizationHeader).getUserID();
         chatConsumer.leaveChat(chat_id, user_id).block();
@@ -83,11 +83,11 @@ public class ChatResolver {
     }
 
     @MutationMapping
-    public Map<String, String> deleteChat(@ContextValue("Authorization") String AuthorizationHeader, @Argument String chat_id) {
+    public Map<String, String> deleteChat(@ContextValue("Authorization") String AuthorizationHeader, @Argument String chat_id) throws Exception {
 
-        // if (!authConsumer.CheckRoute("/chat/chat:id", "delete", AuthorizationHeader)) {
-        //     throw new Exception("Auth Problem");
-        // }
+        if (!authConsumer.CheckRoute("/chat/chat:id", "delete", AuthorizationHeader)) {
+            throw new Exception("Auth Problem");
+        }
 
         chatConsumer.deleteChat(chat_id).block();
         Map<String, String> response = new HashMap<>();
